--- assets/build-configs/settings/vite.config.ts
@@ -1,18 +1,0 @@
-import type {UserConfig} from 'vite';
-import path from 'path';
-import {makeViteConfig} from "../vite.config";
-
-const settingsConfig: UserConfig = {
-    build: {
-        rollupOptions: {
-            input: {
-                settings: path.resolve(__dirname, '../../src/settings/settings.scss'),
-            },
-        },
-    },
-    css: {
-        postcss: path.resolve(__dirname, './postcss.config.js'),
-    },
-};
-
-export default makeViteConfig("settings", settingsConfig);
