--- assets/build-configs/statistics/postcss.config.js
@@ -1,10 +1,0 @@
-const path = require('path');
-
-module.exports = {
-    plugins: {
-        tailwindcss: {
-            config: path.resolve(__dirname, './tailwind.config.js')
-        },
-        autoprefixer: {},
-    },
-};
