--- assets/build-configs/tailwind-theme.json
@@ -1,18 +1,0 @@
-{
-  "colors": {
-    "blue": "#2271b1",
-    "blue-dark": "#135e96",
-    "gray": "#8c8f94"
-  },
-  "fontFamily": {
-    "sans": [
-      "ui-sans-serif",
-      "system-ui",
-      "sans-serif",
-      "Apple Color Emoji",
-      "Segoe UI Emoji",
-      "Segoe UI Symbol",
-      "Noto Color Emoji"
-    ]
-  }
-}
