--- assets/build-configs/widget/vite.config.ts
@@ -1,15 +1,0 @@
-import type {UserConfig} from "vite";
-import path from "path";
-import {makeViteConfig} from "../vite.config";
-
-const widgetConfig: UserConfig = {
-    build: {
-        rollupOptions: {
-            input: {
-                widget: path.resolve(__dirname, "../../src/widget/widget.ts"),
-            },
-        },
-    },
-};
-
-export default makeViteConfig("widget", widgetConfig);
