+++ assets/package.json
@@ -1,14 +1,10 @@
 {
   "scripts": {
-    "dev:settings": "vite --config build-configs/settings/vite.config.ts",
-    "dev:statistics": "vite --config build-configs/statistics/vite.config.ts",
-    "dev:widget": "vite --config build-configs/widget/vite.config.ts",
-    "dev:integrations": "vite --config build-configs/integrations/vite.config.ts",
-    "build:settings": "vite build --config build-configs/settings/vite.config.ts",
-    "build:statistics": "vite build --config build-configs/statistics/vite.config.ts",
-    "build:widget": "vite build --config build-configs/widget/vite.config.ts",
-    "build:integrations": "vite build --config build-configs/integrations/vite.config.ts",
-    "build:all": "yarn build:settings && yarn build:statistics && yarn build:widget && yarn build:integrations",
+    "dev:settings": "vite --config src/settings/vite.config.ts",
+    "dev:widget": "vite --config src/widget/vite.config.ts",
+    "build:settings": "vite build --config src/settings/vite.config.ts",
+    "build:widget": "vite build --config src/widget/vite.config.ts",
+    "build:all": "yarn build:settings && yarn build:widget",
     "lint:check": "eslint src",
     "lint:fix": "eslint src --fix",
     "prettier:check": "prettier src --check",
