+++ assets/src/settings/statistics-styles.scss
