+++ assets/src/settings/statistics/box.tsx
