+++ assets/src/settings/statistics/collection.ts
