+++ assets/src/settings/statistics/config.ts
