+++ assets/src/settings/statistics/numberUtils.ts
