+++ assets/src/settings/statistics/usageInfo.tsx
