+++ prosopo-procaptcha/src/Settings/Tabs/Statistics.php
@@ -33,11 +33,11 @@ class Statistics extends Procaptcha_Settings_Tab {
 	}
 
 	public function get_tab_script_asset(): string {
-		return 'statistics/statistics.tsx';
+		return 'settings/statistics.tsx';
 	}
 
 	public function get_style_asset(): string {
-		return 'statistics/statistics-styles.scss';
+		return 'settings/statistics-styles.scss';
 	}
 
 	public function get_tab_js_data( Settings_Storage $settings_storage ): array {
